split out db
